+++ src/ZoneTree/Segments/Disk/FileRandomAccessDevice.cs
@@ -13,7 +13,7 @@ public sealed class FileRandomAccessDevice : IRandomAccessDevice
     readonly IRandomAccessDeviceManager RandomDeviceManager;
 
     public string FilePath { get; }
-    public int SegmentId { get; }
+    public long SegmentId { get; }
 
     public bool Writable { get; }
